+++ test_settings.py
@@ -1,0 +1,54 @@
+"""Minimal Django settings for running brevo_analytics tests standalone."""
+
+SECRET_KEY = 'test-secret-key-for-brevo-analytics'
+
+DATABASES = {
+    'default': {
+        'ENGINE': 'django.db.backends.sqlite3',
+        'NAME': ':memory:',
+    }
+}
+
+INSTALLED_APPS = [
+    'django.contrib.contenttypes',
+    'django.contrib.auth',
+    'django.contrib.sessions',
+    'django.contrib.messages',
+    'django.contrib.admin',
+    'rest_framework',
+    'corsheaders',
+    'brevo_analytics',
+]
+
+MIDDLEWARE = [
+    'django.contrib.sessions.middleware.SessionMiddleware',
+    'django.middleware.common.CommonMiddleware',
+    'django.middleware.csrf.CsrfViewMiddleware',
+    'django.contrib.auth.middleware.AuthenticationMiddleware',
+    'django.contrib.messages.middleware.MessageMiddleware',
+]
+
+TEMPLATES = [
+    {
+        'BACKEND': 'django.template.backends.django.DjangoTemplates',
+        'DIRS': [],
+        'APP_DIRS': True,
+        'OPTIONS': {
+            'context_processors': [
+                'django.template.context_processors.debug',
+                'django.template.context_processors.request',
+                'django.contrib.auth.context_processors.auth',
+                'django.contrib.messages.context_processors.messages',
+            ],
+        },
+    },
+]
+
+ROOT_URLCONF = 'test_urls'
+
+BREVO_ANALYTICS = {
+    'ALLOWED_SENDERS': [],
+    'EXCLUDED_RECIPIENT_DOMAINS': [],
+}
+
+DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'